Fix putShort to accept Int; add all-zero/all-ones layout tests; add putFloat/putDouble and SnapshotSerializer format tests
Agent-Logs-Url: https://github.com/pjfanning/incubator-pekko/sessions/3a10e6a7-4e61-4f22-9ca1-fdbad5891385

+++ actor/src/main/scala/org/apache/pekko/util/ByteString.scala
@@ -2246,7 +2246,7 @@ final class ByteStringBuilder extends Builder[Byte, ByteString] {
    */
   def putShort(x: Int)(implicit byteOrder: ByteOrder): this.type = {
     ensureTempSize(_tempLength + 2)
-    SWARUtil.putShort(_temp, _tempLength, x.toShort, byteOrder)
+    SWARUtil.putShort(_temp, _tempLength, x, byteOrder)
     _tempLength += 2
     _length += 2
     this
